Rename datafeed to vatsim
- This makes more sense, since - f.e. - the status file is also queried from within this module.
